refactor of Tk_info
