+++ function_data_conversion/main.py
@@ -1,20 +1,22 @@
+import json
 import traceback
-import google.cloud.storage
+from os import remove
 from pathlib import Path
 from typing import Union
-from google.cloud.storage.blob import Blob
+
+import google.cloud.storage
 from flask.wrappers import Request
 from fuzzywuzzy import fuzz
-from os import remove
-import json
+from google.cloud.storage.blob import Blob
+
 from converter import Convert_lines_to_army_list, Write_army_lists_to_json_file
-from new_recruit_tournaments import armies_from_NR_tournament
-from game_report import armies_from_report
 from fading_flame import armies_from_fading_flame
-from warhall import armies_from_warhall
-from tourney_keeper import get_recent_tournaments
-from utility_functions import Docx_to_line_list
+from game_report import armies_from_report
 from multi_error import Multi_Error
+from new_recruit_tournaments import armies_from_NR_tournament
+from tourney_keeper import get_recent_tournaments, armies_from_docx
+from utility_functions import Docx_to_line_list
+from warhall import armies_from_warhall
 
 
 def download_blob(bucket_name, blob_name) -> Union[Blob, None]:
@@ -71,7 +73,7 @@ def function_data_conversion(request: Request) -> tuple[dict, int]:
                 event_name = Path(download_file_path).stem
                 lines = Docx_to_line_list(download_file_path)
 
-                list_of_armies = Convert_lines_to_army_list(event_name, lines)
+                list_of_armies = armies_from_docx(event_name, lines)
             except Multi_Error as e:
                 print(f"Multi_ErrorWD: {[str(x) for x in e.errors]}")
                 return {"message": [str(x) for x in e.errors]}, 400
@@ -138,10 +140,10 @@ def function_data_conversion(request: Request) -> tuple[dict, int]:
             list_of_armies = armies_from_warhall(data)
 
         except Multi_Error as e:
-            print(f"Multi_ErrorWH: {[str(x) for x in e.errors]}")
+            print(f"File Name: {file_name}, Multi_ErrorWH: {[str(x) for x in e.errors]}")
             return {"message": [str(x) for x in e.errors]}, 400
         except Exception as e:
-            print(f"Non Multi ErrorWH: {str(type(e))}, {str(e)}")
+            print(f"File Name: {file_name}, Non Multi ErrorWH: {str(type(e))}, {str(e)}")
             return {"message": [str(e)]}, 400
 
     # NEW RECRUIT TOURNAMENTS
@@ -162,10 +164,10 @@ def function_data_conversion(request: Request) -> tuple[dict, int]:
             list_of_armies = armies_from_NR_tournament(data)
 
         except Multi_Error as e:
-            print(f"Multi_ErrorNR: {[str(x) for x in e.errors]}")
+            print(f"File Name: {file_name}, Multi_ErrorNR: {[str(x) for x in e.errors]}")
             return {"message": [str(x) for x in e.errors]}, 400
         except Exception as e:
-            print(f"Non Multi ErrorNR: {str(type(e))}, {str(e)}")
+            print(f"File Name: {file_name}, Non Multi ErrorNR: {str(type(e))}, {str(e)}")
             return {"message": [str(e)]}, 400
 
     else:
@@ -234,7 +236,7 @@ if __name__ == "__main__":
     # json_message = {"data": {'deployment_selected': ['2 Dawn Assault'], 'dropped_all': ['player2'], 'game_date': ['2022-01-06'], 'map_selected': ['A8'], 'name': 'manual game report', 'objective_selected': ['1 Hold the Ground'], 'player1_army': ["Warriors of the Dark Gods\r\n485 - Sorcerer, Wizard Master, Evocation, Light Armour, Magical Heirloom, Ranger's Boots\r\n340 - Barbarian Chief, General, Black Steed (Prized Stallion), Shield, Heavy Armour, Hand Weapon (Burning Portent), Potion of Swiftness\r\n295 - Barbarian Chief, Shadow Chaser, Heavy Armour (Thrice-Forged), Paired Weapons (Shield Breaker)\r\n235 - Barbarian Chief, War Dais, Heavy Armour, Paired Weapons (Symbol of Slaughter), Rod of Battle\r\n385 - 40 Barbarians, Throwing Weapons, Paired Weapons, Standard Bearer (Wasteland Torch), Musician, Champion\r\n295 - 30 Barbarians, Shield, Standard Bearer (Legion Standard), Musician, Champion\r\n220 - 10 Fallen\r\n520 - 8 Warrior Knights, Lance, Pride, Standard Bearer (Stalker's Standard), Musician, Champion\r\n409 - 3 Chosen Knights, Wrath\r\n330 - 4 Wretched Ones\r\n320 - Battleshrine\r\n300 - Marauding Giant, Tribal Warspear\r\n200 - Chimera\r\n165 - 5 Flayers, Shield\r\n4499\r\n"], 'player1_magic': ['H', 'E1', 'E2', 'E3', 'E5', 'O6'], 'player1_name': ['Tom'], 'player1_score': ['16'], 'player1_vps': [''], 'player2_army': ["Kingdom of Equitaine\r\n505 - Equitan Lord, Pegasus Charger, Shield, Lance (Divine Judgement), Basalt Infusion, Sacred Chalice, Paladin, Honour\r\n485 - Damsel, Destrier, Wizard Master, Shamanism, Binding Scroll\r\n395 - Damsel, General, Revered Unicorn, Wizard Adept, Divination, Crystal Ball, Sainted\r\n340 - Folk Hero, Destrier, Paired Weapons, Battle Standard Bearer (Aether Icon, Aether Icon), Light Armour (Essence of Mithril), Bannerman, Castellan, Faith\r\n670 - 15 Feudal Knights, Champion (Knight Banneret), Musician, Standard Bearer (Stalker's Standard)\r\n275 - 6 Feudal Knights, Champion, Musician\r\n275 - 6 Feudal Knights, Champion, Musician\r\n650 - 5 Pegasus Knights, Champion (Knight Banneret (Oriflamme)), Standard Bearer (Banner of Speed)\r\n340 - 6 Sky Heralds, Paired Weapons, Champion\r\n135 - 5 Yeoman Outriders, Bow\r\n430 - The Lady's Courtier, Courtier of the Dawn\r\n4500"], 'player2_magic': ['H', 'DV3', 'DV4', 'S1', 'S2', 'S6'], 'player2_name': ['Erdem'], 'player2_score': ['4'], 'player2_vps': [''], 'who_deployed': ['player2'], 'who_started': ['player2'], 'won_secondary': ['player1']}}
     # json_message = {'data': {'deployment_selected': ['2 Dawn Assault'], 'game_date': ['2022-02-25'], 'map_selected': ['A1'], 'name': 'manual_game_report', 'objective_selected': ['6 Secure Target'], 'player1_army': ['810 tengu commander general alpha carnosaur halberd (deathstalker) obsidian rock\r\n155 skink commander master tracker infiltrator chart\r\n815 anurarch guardian divination master tablet of atua binding scroll spell breaker mindreader copycat forbidenpath\r\n\r\n380 19 tengu spear pattern of the clever girl\r\n250 15 tengu\r\n205 20 skink javelineers\r\n205 20 skink javelineers\r\n189 9 raptor packs\r\n\r\n300 2X5 cameleons\r\n360 2X3 pteradons\r\n290 2X5 skink hunter with pheromon\r\n\r\n540 1 taurosaur BSB\r\n\r\ntotal 4499'], 'player1_magic': ['P3'], 'player1_name': ['lamronchak'], 'player1_score': ['11'], 'player1_vps': ['2313'], 'player2_army': ["Undying Dynasties\r\n485 - Death Cult Hierarch, Wizard Master, Cosmology, Binding Scroll, Talisman of the Void\r\n300 - Death Cult Hierarch, Wizard Adept, Evocation, Book of Arcane Mastery, Hierophant\r\n430 - Pharaoh, General, Shield (Sun's Embrace), Light Armour (Destiny's Call), Hand Weapon (Scourge of Kings), Death Mask of Teput\r\n200 - Tomb Architect, Light Armour\r\n517 - 59 Skeletons, Spear, Standard Bearer (Aether Icon), Musician, Champion\r\n195 - 17 Skeleton Archers, Musician\r\n125 - 10 Skeleton Archers, Musician\r\n145 - 5 Skeleton Scouts\r\n145 - 5 Skeleton Scouts\r\n712 - 8 Shabtis, Paired Weapons, Musician\r\n525 - 5 Tomb Cataphracts, Musician\r\n291 - 4 Sand Stalkers\r\n430 - Dread Sphinx\r\n4500"], 'player2_magic': ['W5'], 'player2_name': ['gundizalbo'], 'player2_score': ['9'], 'player2_vps': ['1625'], 'who_deployed': ['player1'], 'who_started': ['player2']}}
     json_message = {"data": {
-  "event_id": "626ed0a51508e850435e319f",
+  "event_id": "62947cf2707f76f796db67ed",
   "name": "newrecruit_tournament.json"
 }}
     request_obj = Request.from_values(json=json_message)
