+++ function_data_conversion/tourney_keeper.py
@@ -8,9 +8,11 @@ from unicodedata import category
 from unidecode import unidecode
 from uuid import UUID, uuid4
 from fuzzywuzzy import fuzz
-from data_classes import ArmyEntry, Tk_info, Event_types, Data_sources, Round
+from data_classes import ArmyEntry, Army_names, Tk_info, Event_types, Data_sources, Round
 from functools import cache
 import re
+from converter import Convert_lines_to_army_list
+
 
 http = requests.Session()
 
@@ -164,7 +166,7 @@ def Get_players_names_from_games(games: dict) -> dict:
         Key = tk_player_id
         dict: {1234: {TournamentPlayerId: 5678, Player_name: bob}}
     """
-    # interate over all games and produce list of unique player ids
+    # iterate over all games and produce list of unique player ids
     unique_player_tkIds = set()
     for game in games:
         unique_player_tkIds.add(game.get("Player1Id"))
@@ -279,8 +281,6 @@ def append_tk_game_data(
             player1 = next( (x for x in tk_info.player_list.values() if x.get("TournamentPlayerId") == game.get("Player1Id")), {})
             player2 = next( (x for x in tk_info.player_list.values() if x.get("TournamentPlayerId") == game.get("Player2Id")), {})
 
-            
-
             # Check that non active players have lists. If not skip this game data since only 1 player from the round will be recorded and so averages will be thrown off.
             if not player1.get("Active") and not any(x.player_name == player1.get("Player_name") for x in list_of_armies):
                 continue
@@ -318,6 +318,8 @@ def append_tk_game_data(
 
             for army in list_of_armies:
                 army.data_source = Data_sources.TOURNEY_KEEPER
+                army.event_date = tk_info.event_date
+                army.event_type = tk_info.event_type
                 if army.army_uuid == player1_uuid:
                     if not army.round_performance:
                         army.round_performance = []
@@ -351,3 +353,178 @@ def append_tk_game_data(
             )
         for index, army in enumerate(list_of_armies):
             army.list_placing = index + 1  # have to account for 0 index lists
+
+def match_player_to_tk_name(tk_info: Tk_info, army: ArmyEntry) -> None:
+    """
+    Match the player name to the TK name
+    """
+    if tk_info.player_list:
+        # fuzzy match name from lists file and tourney keeper
+        close_matches = [
+            (
+                item,
+                ratio,
+            )
+            for item in tk_info.player_list.items()
+            if (
+                ratio := fuzz.token_sort_ratio(item[1]["Player_name"], army.player_name)
+            )
+            > 50
+        ]
+        if len(close_matches) > 0:
+            sorted_by_fuzz_ratio = sorted(
+                close_matches, key=lambda tup: tup[1], reverse=True
+            )
+
+            if (
+                sorted_by_fuzz_ratio[0][1] != 100
+            ):  # only report when there are a few options and the top pick isn't 100
+                raise ValueError(
+                    f"No perfect matches for '{army.player_name}' in {sorted_by_fuzz_ratio}"
+                )
+
+            top_picks = [x for x in sorted_by_fuzz_ratio if x[1] == 100]
+            if len(top_picks) > 1:
+                # reduce number of top picks down by army played
+                top_picks = [
+                    x
+                    for x in top_picks
+                    if Army_names[x[0][1].get("Primary_Codex", army.army).upper()] == army.army
+                ]
+
+                if len(top_picks) > 1:
+                    raise ValueError(
+                        f"These {len(top_picks)} players are indistinguishable: {top_picks}"
+                    )
+                elif len(top_picks) == 0:
+                    raise ValueError(
+                        f"None of the tk matches for {army.player_name} played {army.army} as found in the word docx."
+                    )
+            # set current army to be the top pick
+            army.tourney_keeper_TournamentPlayerId = top_picks[0][0][1].get(
+                "TournamentPlayerId"
+            )
+            army.tourney_keeper_PlayerId = top_picks[0][0][0]
+
+        else:
+            extra_info = "\n".join(army.list_as_str or "HELP")
+            raise ValueError(
+                f"""Player: "{army.player_name}" not on TK
+                Extra info: {extra_info}"""
+            )
+
+def verify_tk_data(army_list: list[ArmyEntry], tk_info: Tk_info):
+    if army_list:
+
+        zipped = list(
+            zip(
+                *[
+                    (x.player_name, x.tourney_keeper_TournamentPlayerId)
+                    for x in army_list
+                    if x.tourney_keeper_TournamentPlayerId
+                ]
+            )
+        )
+        if len(zipped) == 2:
+            matched_player_names = zipped[0]
+            matched_player_tkids = zipped[1]
+            # check to make sure that all players are uniquely identified in tk
+            if len(set(matched_player_tkids)) != len(matched_player_tkids):
+                double_matches = set(
+                    [
+                        x
+                        for x in matched_player_tkids
+                        if matched_player_tkids.count(x) > 1
+                    ]
+                )
+                doubles_with_name = set(
+                    [
+                        x
+                        for x in zip(matched_player_names, matched_player_tkids)
+                        if x[1] in double_matches
+                    ]
+                )
+
+                raise(
+                    ValueError(
+                        f"""Players duplicated in word file and not uniquely mapped to tk:\n {doubles_with_name}"""
+                    )
+                )
+
+            if (
+                tk_info.player_list
+                and len(tk_info.player_list) != len(matched_player_tkids)
+            ):
+                # If we have the player count from TK then we can check that the number of lists we read in are equal
+                from_file = [x.player_name for x in army_list]
+                from_tk = [x["Player_name"] for x in tk_info.player_list.values()]
+
+                # difference doesn't work here because we are fuzz matching
+                unique_from_file = from_file[:]
+                unique_from_tk = from_tk[:]
+                for x in from_file:
+                    for y in from_tk:
+                        if fuzz.token_sort_ratio(x, y) == 100:
+                            try:
+                                unique_from_file.remove(x)
+                                unique_from_tk.remove(y)
+                            except ValueError:
+                                # This happens when there are 2 player names that are the same and so the value can not be removed.
+                                # This is already handled above with the message of all duplicated players so does not need handling here
+                                pass
+
+                # if all the "missing" tk names are not active then ignore this error
+                if any(missing_actives:=[y.get("Player_name") for x in unique_from_tk for y in tk_info.player_list.values() if y.get("Player_name") == x and y.get("Active")]):
+                    raise(
+                        ValueError(
+                            f"Lists read: {len(army_list)}\nActive players on tourneykeeper: {tk_info.player_count}\nPlayers matched: {len(matched_player_tkids)}\nPlayers in file but not TK: {unique_from_file}\nPlayers in TK but not in file: {missing_actives}"
+                        )
+                    )
+        # If we have tk data but zipping player id's failed
+        elif tk_info.event_id:
+            raise(ValueError(f"No tkdata was loaded into armies"))
+    
+
+def armies_from_docx(event_name: str, lines: list[str]) -> List[ArmyEntry]:
+    armies = Convert_lines_to_army_list(event_name=event_name, lines=lines)
+    tk_info = load_tk_info(event_name)
+    if tk_info and tk_info.game_list and tk_info.player_list: #game was found on tk
+        for army in armies:
+            match_player_to_tk_name(tk_info=tk_info, army=army)
+        append_tk_game_data(tk_info=tk_info, list_of_armies=armies)
+        verify_tk_data(army_list=armies, tk_info=tk_info)
+    return armies
+
+if __name__ == "__main__":
+    """Used for testing locally"""
+    import os
+    from time import perf_counter
+    from utility_functions import Docx_to_line_list
+    from pathlib import Path
+
+
+    t1_start = perf_counter()
+
+    path = Path("../data/list-files")
+
+    os.makedirs(os.path.dirname(path / "json"), exist_ok=True)
+    for file in os.listdir(path):
+        if file.endswith(".docx") and not file.startswith("~$"):
+            file_start = perf_counter()
+            filePath = Path(os.path.join(path, file))
+            event_name = Path(filePath).stem
+            print(f"Input filepath = {filePath}")
+
+            lines = Docx_to_line_list(filePath)
+            list_of_armies = armies_from_docx(event_name, lines)
+            file_stop = perf_counter()
+            print(
+                f"{len(list_of_armies)} army lists were found in {round(file_stop - file_start)} seconds"
+            )
+            print(f"Player Name list: {[army.player_name for army in list_of_armies]}")
+            if all(x.tourney_keeper_PlayerId for x in list_of_armies):
+                print(f"Tk Info loaded")
+            else:
+                print(f"TK Not loaded")
+    t1_stop = perf_counter()
+    print(f"Total Elapsed time: {round(t1_stop - t1_start)} seconds")
